refactor: extract fallback-selection logic into model-fallback.js
Split model fallback selection out of model-resolver.js (431 lines) into a
dedicated model-fallback.js module (~130 lines):

Extracted functions:
- normalizeFallbackConfig
- resolveAliasDefinition
- inferModelFamilyPrefix
- selectMiddlePowerFallback
- tryMiddlePowerFallback

model-resolver.js imports from model-fallback.js and continues to re-export
selectMiddlePowerFallback for downstream callers.

Closes #5632
