+++ containers/api-proxy/model-resolver.js
@@ -21,6 +21,14 @@
 
 const { getTierSortedModels } = require('./model-discovery');
 const { globMatch, extractVersionNumbers, compareByVersion } = require('./model-utils');
+const {
+  DEFAULT_MODEL_FALLBACK,
+  normalizeFallbackConfig,
+  resolveAliasDefinition,
+  inferModelFamilyPrefix,
+  selectMiddlePowerFallback,
+  tryMiddlePowerFallback,
+} = require('./model-fallback');
 
 /**
  * Check whether a model name is permitted by the given policy config.
@@ -40,11 +48,6 @@ function _isModelPermittedByPolicy(model, policyConfig) {
   if (allowedModels && !allowedModels.some(p => globMatch(p, model))) return false;
   return true;
 }
-
-const DEFAULT_MODEL_FALLBACK = Object.freeze({
-  enabled: true,
-  strategy: 'middle_power',
-});
 
 /**
  * Parse model aliases configuration from a raw JSON string.
@@ -84,82 +87,6 @@ function parseModelAliases(rawConfig) {
   }
 
   return { models: parsed.models };
-}
-
-function normalizeFallbackConfig(modelFallbackConfig) {
-  const config = modelFallbackConfig && typeof modelFallbackConfig === 'object'
-    ? modelFallbackConfig
-    : DEFAULT_MODEL_FALLBACK;
-  return {
-    enabled: config.enabled !== false,
-    strategy: config.strategy || 'middle_power',
-  };
-}
-
-function resolveAliasDefinition(rawAlias) {
-  if (Array.isArray(rawAlias)) {
-    return { patterns: rawAlias, fallback: true };
-  }
-  if (!rawAlias || typeof rawAlias !== 'object' || Array.isArray(rawAlias)) {
-    return { patterns: [], fallback: true };
-  }
-  return {
-    patterns: Array.isArray(rawAlias.patterns) ? rawAlias.patterns : [],
-    fallback: rawAlias.fallback !== false,
-  };
-}
-
-function inferModelFamilyPrefix(requestedModel) {
-  const key = String(requestedModel || '').toLowerCase();
-  const gptFamily = key.match(/^(gpt-\d+(?:\.\d+)?)/)?.[1];
-  if (gptFamily) return gptFamily;
-  if (key.includes('claude')) return 'claude';
-  if (key.includes('gemini')) return 'gemini';
-  return null;
-}
-
-function selectMiddlePowerFallback(requestedModel, availableModels, currentProvider, reason, modelFallbackConfig) {
-  const fallbackConfig = normalizeFallbackConfig(modelFallbackConfig);
-  if (!fallbackConfig.enabled || fallbackConfig.strategy !== 'middle_power') return null;
-
-  const providerModels = Array.isArray(availableModels[currentProvider]) ? availableModels[currentProvider] : [];
-  if (providerModels.length === 0) return null;
-
-  const familyPrefix = inferModelFamilyPrefix(requestedModel);
-  const familyCandidates = familyPrefix
-    ? providerModels.filter(model => model.toLowerCase().startsWith(familyPrefix))
-    : [];
-  const selectedPool = familyCandidates.length > 0 ? familyCandidates : providerModels;
-  const sortedCandidates = getTierSortedModels(currentProvider, selectedPool);
-  if (sortedCandidates.length === 0) return null;
-
-  const medianIndex = Math.floor((sortedCandidates.length - 1) / 2);
-  return {
-    resolvedModel: sortedCandidates[medianIndex].model,
-    fallback: {
-      activated: true,
-      reason,
-      selection_method: 'middle_power_median',
-      available_models_count: providerModels.length,
-      used_family_filter: familyCandidates.length > 0,
-      candidates: sortedCandidates,
-    },
-  };
-}
-
-/**
- * Attempts middle-power fallback and returns a resolution result if successful.
- * Encapsulates the repeated call + log + return pattern used in two places.
- */
-function tryMiddlePowerFallback(requestedModel, availableModels, currentProvider, reason, fallbackConfig, log) {
-  const middlePowerFallback = selectMiddlePowerFallback(
-    requestedModel, availableModels, currentProvider, reason, fallbackConfig
-  );
-  if (middlePowerFallback) {
-    log.push(`[model-resolver] middle-power fallback: "${requestedModel}" → "${middlePowerFallback.resolvedModel}"`);
-    return { resolvedModel: middlePowerFallback.resolvedModel, log, fallback: middlePowerFallback.fallback };
-  }
-  return null;
 }
 
 /**
